Move custom pact Given extensions to PactBuilderExtensions.cs
